(SZKOPUL-DW) Fix the get_report_HTML TypeError again
I think the problem was in the translation laziness all along. I am
inclined to believe we don't need the laziness here, because we are
already in a request.

Steps to repro:

 - make a new manual deployment
 - launch supervisor
 - ./manage.py supervisor stop oioioiworker
 - submit something as user
 - go to http://localhost:8000/c/yourcontestname/submissions/
 - tap your submission (not the link)

Before this patch: AJAX gets a 500 error, silently shows "No reports
are available" (translated)

After this patch: AJAX gets a 200, shows the new cool and cute message

Change-Id: I8beb0a1e69188c95484c8133286243e73bfa93f4