(SENTRY-something) Fix mixed strings in package processing status
This little case of mixing up Unicode strings with bytestrings in
formatting gave a beautiful UnicodeDecodeError or something like that,
making between 50 and 150 uploaded packages hang indefinitely as
awaiting processing because they had a compilation error and as it tried
to bubble up from the workers it had an unexpected exception.

I'm not sure this is a definitive solution, but hey, it's Python, what
can you do.

Change-Id: I8f207076d2ecb91779ed5ef977a3072701e354fb