SIO-2029 Fix shift+enter copying in test editing view
When editing test limits and scores in a probleminstance,
shift+enter is supposed to copy the value from the current field
into all subsequent fields in the same column.

It broke again, when the inputs' type changed to number,
which means jQuery's input:text selector no longer finds them.
Fix this by explicitly matching the right type of input.

Change-Id: If3dc801547a00ddfd75753a93c8a1d3d5733b706