Handle list objects in quiz's create_submission
When creating a quiz submission not from something else
than real post data (eg. from a management command)
the form_data may be a dict of lists, which behaves differently
for text-input questions - it returns a one-element list instead
of a single value.

Adapt QuizProblemController to handle that as well.

Change-Id: Ib84da7bc581c746e139915e2ee8b1f0e6d4191f5