(no-ticket) Fix 500 errors on missing statements
The views which handle statements as binary files (zips, pdfs, etc)
didn't check if query_statement returns None, which it does when
there is no problem statement.

Make them check for that, and return 404 instead of crashing.

Links to these views shouldn't appear to regular users for
problems without statements, and these views aren't supposed
to return HTML, so we don't show the "no problem statement" message
like we do in other cases.

Change-Id: I60a503691e818b8fa61baf24d807b8a3acc5eb96