SIO-2297 Fixed bug with rejudging old submissions after adding new questions to a quiz
I modified method responsible for scoring quiz, as well as method for scoring
question so that it is checked if any answer was submitted for a particular question.
If there is no submitted answer, then such question is ignored and not counted to
the overall quiz score.
I added two test cases for rejudging quiz after adding new questions to it.

Change-Id: Idc4c0452b6bce4494b050dcdb22fa37a261d7820