(no-ticket) Fix race condition in package upload API
This is the same race that 7ff5a3 fixed in add_or_update_problem view,
but this time it's in API views.

We were creating an unpackmgr job within the implicit view-wide
database transaction, which allowed it to start executing before
commit, so unpackmgr didn't see any of our database changes.

Solve this by disabling the implicit view-wide transaction
on the package (re)upload API views, since process_package
makes the right size of a transaction on its own.

Another solution would be to defer the unpackmgr job
to be enqueued in transaction.on_commit,
but it's simpler to stay consistent with the fix in 7ff5a3.

Change-Id: I94e3552454ca84ee7a1d147d8e69da2d426217c8