SIO-2367 Add submit button in the contest's problems list to simplify the solution submit process
When a user is in the Problems submenu and wants to upload a solution, he has to go to the Submit submenu and manually select his problem.
This commit adds the submit button to the problems list, which redirects users to the Submit submenu with the chosen problem already selected for them.

Change-Id: Ia5e49731b817d707ee8f21a26d13cc6274b0d3e8