(SIO-2399) Update static analysis config files
Adds new configs for black and isort and updates configs for pep8 and
pylint.
Adds .git-blame-ignore-revs file that containts ref numbers of commits
that should be ignored by git blame.

It may be necessary to run
git config blame.ignoreRevsFile .git-blame-ignore-revs
for git to always respect this config when running blame.

Change-Id: I9645a2d9a8772e92bf99ce9fbaa6a98a0c94b422