SIO-2048 Allow customizing the disq message
Part of the message participant sees is fixed. This part says that
participant is disqualified and can use complaint forms to plea. This is
not accurate for some competitions and so there is need for ability to
customize the message.

New part to "Advanced" section of contest config is added, where the
message can be customized. Disqualification mixins ask for this custom
message and give those to templates renderers.

Change-Id: I4851f2f052b517d10977df2aed9618e56a090d15