(SIO-2348) Add a note that tests are public to Add to contest tab.
The teacher while adding the problem to their contest should
be aware of possible cheating -- and downloading tests from problemset
may be a way to do it. In add to contest tab there is a big warning
if tests package is public now. I also updated the README to make
conditions of tests publishing clearer.

Change-Id: Ief823e09e176b96481c1debb07d2d25cce9f6074