SIO-2409 Change user choice widget in contest permissions admin
The default fk object selection widget fetches all possible choices from
the db and creates a dropdown select with options. This caused problems
because with 10s of thousands of users the browser would freeze when
loading the options.

This commit changes the widget. Now the field accepts just a plaintext
username.

Change-Id: I4d6196a94c35a8100cf32acfacd8ab203dd9923c