(no-ticket) Algorithm and origin tags searching fixes
There is a search bar for the users to search for interesting
problems by their names and attached tags in the problemset.
Apparently, there was a problem with searching for algorithm
tags and origin tags. For the first, translated names in
the currently picked language were not properly converted
to values used by client-side code for searching for problems.
For the latter, there was a problem with reloading menu
with prompts if clicked origin tag's full name was a substring
of another origin tag's full name.

*Use an algorithm tag's name as a value used to search for
 problems, instead of an algorithm tag's full name.
*Create the new view returning only clicked origin tag and its
 categories.
*Modify and refactor exisiting JavaScript code to accomodate for
 the new view.
*Refactor old Python code.
*Add tests for the new view.

Change-Id: I84a1732a6ac7a82146a0f4233a6abc8c9cc5fa54