SIO-2055 Add support for Output-only
Adds Output-only to keys in settings. Introduces new meta data key for
every language: "type". This key can be set to "main" or "extra", the
later means that the language can only be enabled by listing it on
language whitelist for certain problem. An only language of this type is
"Output-only", as it would probably be used quite rarely and enabling it
on every problem by default might confuse some users (especially
considering fact, that problem has to be prepared in special way in
order to be solvable by output-only).

Change-Id: I284b16c9bbbd7af021827abea5f08fedbf4f57f5