(no-ticket) Fix missing submit id in evalqueue admin
In case of non-contests submissions (i.e. problemset ones),
submit_id wasn't displayed in evaluation queue admin.
This was due to an attempt to include contest id in url parameters,
which is needed for contest submissions, but for non-contest ones
passing None as contest_id is fine. Which is what we do now,
by using the _require_contest annotation on _get_contest_id
instead of submit_id.

Change-Id: I2d4efc36f302b1358ff87867e6804df3aed6ba52