SIO-2427 Add django 2.2 support
Removed python_2_unicode_compatible decorator.
Silenced admin.E130 system check.
Removed use of None as url arguemnt.
Started creating object if its id is used in tests as a foreign key.
Adjusted dependencies.
Django 2.2.21 support is not there as it introduced breaking change(name of a file cannot be path) that I'm not sure how to adjust to.

Change-Id: I8858ff4c6ebd69f2271150b11291ed9404ffa372