(no-ticket) Command for creating names' translations of problems
Some of the problems in the Szkopul are repeated: there exist one
version of a particular problem with legacy name in Polish and the
other one with legacy name in English. All of the pairs of repeated
problems share the same short name and exactly one legacy origin tag.
After the change "(SIO-2430) Make problems' names translatable" it
is possible to make problem names translatable by the means of addding
ProblemName objects to a particular problem.

*Add command for creating translations of legacy names of repeated
 problems in Polish and English and associating these translations with
 the problems.
*Add files with all possible legacy origin tags present in the database
 at the production.
*Add tests for the new command.

Change-Id: I6bb9c5fb6b67c4094d285876d7a6d3bb98ead7f3