(no-ticket) Fix reading config.yml with local chars (sinolpack)
When loading config.yml (during processing a sinolpack) with local
characters encoded in utf-8, an error "yaml.reader.ReaderError:
unacceptable character: special characters are not allowed" occurs
because of misinterpretation of bytes encoded with utf-8.
Surprisingly, it does not happen on my local machine, but only on
Jenkins, only when running Python 3 tests.

*Replace standard built-in open() function with io.open() and set
 explicit utf-8 encoding for reading content of the config.yaml
 from the given sinolpack.
*Add a test for unpacking sinolpack with all possible Polish local
 characters present in the config.yml file from that package.
*Rename old tests and corresponding files for better developer's
 experience.

Change-Id: I2c26b4a06915ad35efd998665b177e405b71aba9