(SIO-2396) API: Allow finding a problem by contest_id and short_name
Contest id and problem short name already uniquely identify a problem
instance, so they could be used to find it, and the problem.
It would be nice to have an API that allows obtaining problem_id
(and possibly problem_instance_id) by providing contest id and problem
short name. The problem_id would be very useful for reuploading a package
from a CI, which does not necessarily know it beforehend, but does have
access the contest id and short name in the package source.

*Add an API endpoint for finding problem_id and problem_instance_id
 based on contest_id and problem_short_name.
*Refactor old code.
*Add tests.

Change-Id: Icd4beca8f3668e6a69ee29b1a2e5036293328221