SIO-2457 Bardziej intuicyjny zielony pasek
Until now, the progress bar was indicating how much time has passed,
while the clock was showing how much time there is left.
Now, the progress bar starts full and green and as the time passes
it shrinks down and the color changes appropriately - it shows
how much time there is left until the end of the round.

There had to be a small change in tests made.
One test group checked the count of '%' symbol in the response.
Needed to be incremented by one after my changes.

Change-Id: I9e9308bf0ed72567253d237b1d2ac9bb47716a8d