Fix displaying attachments hidden to non-admins
This is a follow-up of #233, by which only testspackages were handled
somewhat properly. The 'admin_only' field has been added, as the pub_date
alone isn't enough to determine whether the attachment would be visible if
the user wasn't an admin - one example is a ContestAttachment for a round
that starts in the future.