Add problem package upload option to mass_create_tool
Previously, the mass_create_tool only created empty problems.

This commit changes that by adding a list of problem_packages (-pp)
as an option to the tool. The tool searches the oioioi/sinolpack/files/ directory
for the specified problem packages and uploads them to the created contest.

Example command enabled by this commit:

    python3 manage.py mass_create_tool -cn=quite_useful_contest -cc -pp=test_full_package.tgz -u=10 --wipe

    ^ (This will create a contest named 'quite_useful_contest' with 10 users,
       and upload 'test_full_package.tgz' to a new round in that contest.)